+++ Doraboda/app/src/main/AndroidManifest.xml
@@ -14,10 +14,10 @@
         android:theme="@style/Theme.Doraboda"
         tools:targetApi="34">
         <activity
-            android:name=".SettingsActivity"
+            android:name=".view.SettingsActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".view.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
