--- Doraboda/app/src/main/java/com/univ/doraboda/SettingsActivity.kt
@@ -1,11 +1,0 @@
-package com.univ.doraboda
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class SettingsActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_settings)
-    }
-}
