+++ Doraboda/app/src/main/java/com/univ/doraboda/view/DataFragment.kt
@@ -1,0 +1,28 @@
+package com.univ.doraboda.view
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.univ.doraboda.R
+
+/**
+ * A simple [Fragment] subclass.
+ * Use the [DataFragment.newInstance] factory method to
+ * create an instance of this fragment.
+ */
+class DataFragment : Fragment() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_settings, container, false)
+    }
+}
