+++ Doraboda/app/src/main/java/com/univ/doraboda/view/SettingsActivity.kt
@@ -1,0 +1,12 @@
+package com.univ.doraboda.view
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.univ.doraboda.R
+
+class SettingsActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_settings)
+    }
+}
