+++ Doraboda/app/src/main/res/layout/fragment_calendar.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".CalendarFragment">
+    tools:context=".view.CalendarFragment">
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
